requirements-dev.txt
# Development requirements for BusMinder Home Assistant integration

# Testing framework
pytest-homeassistant-custom-component>=0.13.0
aioresponses>=0.7.6
pytest-asyncio>=0.23.0
pytest-cov
